brai: Fix one-handed layouts
For one-handed layouts a question holds on which of the two keys under the
thumb should be dots 5 and 6. Actual tests seem to show that it is more
logical to keep dot 5 next to dot 4. Let's however keep the alternative
as an option.

For left-handed people, it is useful to have a left-hand variant of the
keypad part.

Signed-off-by: Samuel Thibault <samuel.thibault@ens-lyon.org>