Map evdev keycode KEY_ROTATE_LOCK_TOGGLE to XF86RotationLockToggle
Add XF86RotationLockToggle, to be used as mapping for evdev's
KEY_ROTATE_LOCK_TOGGLE.

I've a Point of View P1006W-232 Windows tablet which actually has a
rotate-lock toggle-button. The latest kernel correctly generates
KEY_ROTATE_LOCK_TOGGLE events for this. So now I'm hooking up support for
it through all the higher layers.

The patch for adding the RotationLockToggle keysym to xproto is here:
https://patchwork.freedesktop.org/patch/279365/

Signed-off-by: Hans de Goede <hdegoede@redhat.com>