br: make altgr+e produce degree, not euro sign
It appears that users expect this key to be degree. I did a little bit
of searching, and while I did find a few references to this being euro
sign (e.g., on Wikipedia page [1]), it looks like most photos of
Brazillian keyboards I could find have this as degree, and the user
expectation is that this is degree [2].

For what it's worth, it appears Microsoft Windows has this as
degree (tested on Windows 7).

[1]: https://en.wikipedia.org/wiki/Portuguese_keyboard_layout
[2]: https://crbug.com/298996

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>