Kbd layouts for Mongolian trad. writing (Bichig): Option 1 - China
This patch adds the 7 keyboard layouts included in the Chinese standard
GB/T 28038-2011 to the China (`cn`) symbol file.

The source standard is followed literally, with two exceptions:

1. When the source left the level 2 of a key empty, I have duplicated the
assignment for level 1, so that a 'stray SHIFT' key kept pressed does no
harm.

2. When the standard indicated (mostly for punctuation signs) a vertical
presentation form (UFE30 - UFE4F), I replaced with the original (horizontal)
decomposition, as presentation forms are the job of the font stack rather
than of the keyboard layout and as Mongolian, even in its traditional
script, is nowadays also written horizontally and not only vertically.

This patch inserts the new layouts in the symbol file for China (`cn`), as
Mongolian traditionally writing is mostly used in Chinese Inner Mongolia.
Another, alternative, patch is provided inserting the new layouts in the
Mongolian (`mn`), symbol file on a linguistic, more than geographic, base.