rules: replace the merge.sh with a python script
Easier to read and comprehend, but also safer to use because it's harder to
mix up the headers.

Previously, the HDR file would have the list of section headers and had to be
called in the right order using HDR in the file list in the Makefile.am.

Let's move the section headers to the files themselves where it's more obvious
if they're ever wrong and make the script include them as required.
The include order is kept as-is for now so that the final evdev/base rules
files are identical.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>