rules/compat: switch compat rule generation to a python script
The current generation is a bunch of awk scripts that parse files and spit out
the final rules. This is hard to debug because the knowledge what to parse and
how is in each script (though largely duplicated). Plus, shell scripts are
terrible to read in general.

Move this to a python script instead where debugging is a lot easier and all
the parsing and printing is now in one file. This immediately made obvious
that the special layout/variant 1 handling is just a different base+suffix, so
we can drop that easily.

And let's be more precise in the Makefile too, only give each script the files
it wants.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>