rules: merge sections together into one section per type
We use the Makefile.am order as a guideline in that we keep the same initial
section ordering. So if the sections in the Makefile are A B A C B the outcome
is going to be AA BB C.

This doesn't have any effect right now on the rules file.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>