rules: de-duplicate the parts assignment
There are only a few files that are rule-dependent, the majority of the files
are the same. So let's de-duplicate this here and use a substitute MODEL
The only difference between the base and the evdev parts is that there are
only a few files in each which are rule-dependent. Let's just use a substitute
here which has the nice benefit of making it more obvious where base and evdev
differ.

One file for evdev is now an empty file (the header gets removed during
merge) but we can live with that.

The output is identical where compat rules are enabled, where disabled the new
rules file is now missing a few empty sections - i.e. the functional output is
identical.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>