gitlab CI: run the libxkbcommon xkeyboard-config test tool
This should detect syntax errors and any other parsing issues.

We add a make install stage and pass the installed files through as artifacts
to the next stage. Then we fetch libxkbcommon from git and run the magic
tester tool (which tries to parse every layout/variant/option combination).

A single test run will take quite a while, on my system here that tool takes
about 14 minutes so expect the test suite to be the same.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>