gitlab CI: don't use $HOME in the instdir, it's unset
Let's use CI_PROJECT_DIR which resolves to the root of the checked-out git
directory, i.e. the $PWD for the scripts we use. We can't really use anything
else as install dir because if we want to cache the artifacts (future patch)
they have to be within this directory somewhere.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>