Add cedilla to UK intl. with dead keys
AltGr+, used to map to comma, which is already available by simply
typing comma. Before this change, there was no way to produce a cedilla
from the UK intl. keyboard.

The cedilla is used in popular languages like Portuguese and French. So
much so, that the US intl.  already maps AltGr+, to cedilla. This patch
implements the same feature in the UK intl. keyboad.